requirements.txt
torch
torchvision
timm
matplotlib
numpy
Pillow
PyYAML
